refactor(di): split DataModule into 6 focused Hilt sub-modules
DataModule had grown into an 867-line god module covering networking,
preferences, local stores, streaming/audio, repositories, and sync — every
feature touched the same file and merged into perpetual conflicts. Split
the bindings into domain-grouped sub-modules under di/modules/ and reduce
the root DataModule to a thin @Module(includes = [...]).

Each sub-module is < 200 lines. Pure DI rearrangement — no @Singleton
scope changes and no behaviour change; Hilt KSP graph re-resolves clean
and the JVM test suite stays green.

Closes #342
